+++ examples/data-sources/nxos_vpc_instance/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_vpc_instance" "example" {
+}