+++ examples/resources/nxos_vpc_instance/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_vpc_instance.example "sys/vpc/inst"