+++ examples/resources/nxos_vpc_instance/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_vpc_instance" "example" {
+  admin_state = "enabled"
+}